requirements.txt
tqdm
pydot
numpy
pydot-ng
graphviz
pydotplus
matplotlib
opencv-python
tensorflow-addons
tensorflow-gpu==2.8.0